// backend/src/modules/appointment/enums/time-slot.enum.ts
export enum TimeSlot {
  a0_1 = "a0_1",
  a1_2 = "a1_2",
  a2_3 = "a2_3",
  a3_4 = "a3_4",
  a4_5 = "a4_5",
  a5_6 = "a5_6",
  a6_7 = "a6_7",
  a7_8 = "a7_8",
  a8_9 = "a8_9",
  a9_10 = "a9_10",
  a10_11 = "a10_11",
  a11_12 = "a11_12",
  a12_13 = "a12_13",
  a13_14 = "a13_14",
  a14_15 = "a14_15",
  a15_16 = "a15_16",
  a16_17 = "a16_17",
  a17_18 = "a17_18",
  a18_19 = "a18_19",
  a19_20 = "a19_20",
  a20_21 = "a20_21",
  a21_22 = "a21_22",
  a22_23 = "a22_23",
  a23_24 = "a23_24",
}